MyModelResolver -> MavenModelResolver
